+++ source/reports/train_ml_model_reports/TrainMLModelReport.py
@@ -1,0 +1,15 @@
+from abc import ABC
+
+from source.reports.Report import Report
+
+
+class TrainMLModelReport(Report, ABC):
+    """
+    Train ML model reports plot general statistics or export data of multiple models simultaneously when running the :ref:`TrainMLModel` instruction.
+
+    In the :ref:`TrainMLModel` instruction, train ML model reports can be specified under 'assessment/reports/hyperparameter'.
+    """
+
+    @staticmethod
+    def get_title():
+        return "Train ML model reports"
